ostro-image.bb: always install app-framework by default.
One of the conculsions in the recent QA-related discussions was
that QA wants to be able to run some app-framework tests using
an unmodified image. To enable this, added a dedicated package
group for app-framework and added that to the list of default
package groups to install. This will ensure that the necessary
bits are in the image even without any applications present.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>